Refactor UIManager into TextureManager
