--- Assets/GothicVR/Scripts/Manager/UIManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 34b7e97b277d2b9448536ffb853c106c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
