+++ Assets/GothicVR/Scripts/UI/TextureManager.cs
@@ -4,58 +4,78 @@ using GVR.Caches;
 using GVR.Util;
 using GVR.Phoenix.Interface;
 using TMPro;
-public class TextureManager : MonoBehaviour
+public class TextureManager : SingletonBehaviour<TextureManager>
 {
-    private Texture2D backgroundtexture;
     public Material backgroundmaterial;
-    private Texture2D buttontexture;
     public Material buttonmaterial;
-    private Texture2D slidertexture;
     public Material slidermaterial;
-    private Texture2D sliderpositiontexture;
     public Material sliderpositionmaterial;
-    private Texture2D arrowtexture;
     public Material arrowmaterial;
-    private Texture2D fillertexture;
     public Material fillermaterial;
-    private Texture2D skytexture;
     // public Material skymaterial;
 
-    private bool textureloaded = false;
+    public Material GothicLoadingMenuMaterial;
+    public Material LoadingBarBackgroundMaterial;
+    public Material LoadingBarMaterial;
+    public Material LoadingSphereMaterial;
 
-    void Update()
+    private const string defaultShader = "Universal Render Pipeline/Unlit"; // "Unlit/Transparent Cutout";
+
+    private void Start()
     {
-        // FIXME - We should register to a "BootstrapDone" event rather than checking every frame.
-        if (!textureloaded && GameData.I.VdfsPtr != IntPtr.Zero)
-        {
-            LoadCustomTextures();
-        }
+        GothicLoadingMenuMaterial = GetEmptyMaterial();
+        LoadingBarBackgroundMaterial = GetEmptyMaterial();
+        LoadingBarMaterial = GetEmptyMaterial();
+
+        LoadingSphereMaterial = GetEmptyMaterial();
+        LoadingSphereMaterial.color = new Color(.25f, .25f, .25f, 1f); // dark gray
     }
 
-    public void LoadCustomTextures()
+    public void LoadLoadingDefaultTextures()
     {
-        textureloaded = true;
-        backgroundtexture = AssetCache.I.TryGetTexture("LOG_PAPER.TGA");
+        var loadingBackgroundTexture = AssetCache.I.TryGetTexture("LOADING.TGA");
+        GothicLoadingMenuMaterial.mainTexture = loadingBackgroundTexture;
+
+        var progressBackgroundTexture = AssetCache.I.TryGetTexture("PROGRESS.TGA");
+        LoadingBarBackgroundMaterial.mainTexture = progressBackgroundTexture;
+
+        var progressTexture = AssetCache.I.TryGetTexture("PROGRESS_BAR.TGA");
+        LoadingBarMaterial.mainTexture = progressTexture;
+
+        var backgroundtexture = AssetCache.I.TryGetTexture("LOG_PAPER.TGA");
         backgroundmaterial.mainTexture = backgroundtexture;
 
-        buttontexture = AssetCache.I.TryGetTexture("INV_SLOT.TGA");
+        var buttontexture = AssetCache.I.TryGetTexture("INV_SLOT.TGA");
         buttonmaterial.mainTexture = buttontexture;
 
-        slidertexture = AssetCache.I.TryGetTexture("MENU_SLIDER_BACK.TGA");
+        var slidertexture = AssetCache.I.TryGetTexture("MENU_SLIDER_BACK.TGA");
         slidermaterial.mainTexture = slidertexture;
 
-        sliderpositiontexture = AssetCache.I.TryGetTexture("MENU_SLIDER_POS.TGA");
+        var sliderpositiontexture = AssetCache.I.TryGetTexture("MENU_SLIDER_POS.TGA");
         sliderpositionmaterial.mainTexture = sliderpositiontexture;
 
-        fillertexture = AssetCache.I.TryGetTexture("MENU_BUTTONBACK-C.TEX");
+        var fillertexture = AssetCache.I.TryGetTexture("MENU_BUTTONBACK-C.TEX");
         fillermaterial.mainTexture = fillertexture;
 
-        arrowtexture = AssetCache.I.TryGetTexture("U.TGA");
+        var arrowtexture = AssetCache.I.TryGetTexture("U.TGA");
         arrowmaterial.mainTexture = arrowtexture;
-
-        // FIXME - No material found in Assets. commenting it out right now.
-        // skytexture = AssetCache.I.TryGetTexture("SKYDAY_LAYER0_A0-C.TEX");
-        // skymaterial.mainTexture = skytexture;
     }
 
+    public void SetTexture(string texture, Material material)
+    {
+        material.mainTexture = AssetCache.I.TryGetTexture(texture);
+    }
+
+    private Material GetEmptyMaterial()
+    {
+        var standardShader = Shader.Find(defaultShader);
+        var material = new Material(standardShader);
+
+        material.SetInt("_Cull", (int)UnityEngine.Rendering.CullMode.Off);
+
+        // Enable clipping of alpha values.
+        material.EnableKeyword("_ALPHATEST_ON");
+
+        return material;
+    }
 }
